2nd README update and updated code
Found areas in README that needed improvement. Additionally, uploading currently working code that retrieves information from a microservice. Due to formatting for getting JSON requests, output to the Person Generator GUI includes JSON brackets/formatting. For submission in Sprint 4 this should not be a problem, but will need to be fixed in the future. This formatting does not effect output to .csv